# backend/docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  admin.GamePayload:
    properties:
      category:
        type: string
      description:
        type: string
      icon_url:
        type: string
      key:
        type: string
      name:
        type: string
    required:
    - key
    - name
    type: object
...
info:
  contact: {}
  description: GameLink 平台 API，包含健康检查、认证与管理端能力。
  title: GameLink API
  version: 0.3.0
paths:
  /admin/games:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 列出游戏
      tags:
      - Admin/Games
    post:
      consumes:
      - application/json
      parameters:
      - description: 游戏信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.GamePayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建游戏
      tags:
      - Admin/Games
  /admin/games/{id}:
    delete:
      parameters:
      - description: 游戏ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除游戏
      tags:
      - Admin/Games
    get:
      parameters:
      - description: 游戏ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取游戏
      tags:
      - Admin/Games
    put:
      consumes:
      - application/json
      parameters:
      - description: 游戏ID
        in: path
        name: id
        required: true
        type: integer
      - description: 游戏信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.GamePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新游戏
      tags:
      - Admin/Games
  /admin/games/{id}/logs:
    get:
      parameters:
      - description: 游戏ID
        in: path
        name: id
        required: true
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 动作过滤
        enum:
        - create
        - update
        - delete
        in: query
        name: action
        type: string
      - description: 操作者用户ID
        in: query
        name: actor_user_id
        type: integer
      - description: 开始时间
        in: query
        name: date_from
        type: string
      - description: 结束时间
        in: query
        name: date_to
        type: string
      - description: 导出格式
        enum:
        - csv
        in: query
        name: export
        type: string
      - description: 导出列（逗号分隔）
        in: query
        name: fields
        type: string
      - description: 列头语言
        enum:
        - en
        - zh
        in: query
        name: header_lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取游戏操作日志
      tags:
      - Admin/Games
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"